requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.13
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
openapi-codec==1.3.2
packaging==26.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
simplejson==3.20.2
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
